got the monitor talking

+++ DeltaTPlus_STM32WB55RG_090525/STM32_WPAN/App/p2p_server_app.c
@@ -46,7 +46,7 @@
 
  typedef struct{
     //uint8_t             Device_Button_Selection;
-    uint32_t             DeltaTValue;
+	 float             DeltaTValue;
  }P2P_DeltaTCharValue_t;
 
 typedef struct
